requirements.txt
python-telegram-bot>=20.0
gspread>=5.7.0
google-auth>=2.16.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
